Move functions around eventListeners and triggering flash into util functions.
